app/services/image_service.py
from PIL import Image
import imagehash
import os

def calculate_phash(image_path):
    """Calculates the Perceptual Hash (pHash) of an image."""
    try:
        with Image.open(image_path) as img:
            hash_val = imagehash.phash(img)
            return str(hash_val)
    except Exception as e:
        print(f"Error calculating pHash: {e}")
        return None

def analyze_liveness(image_path):
    """
    Stub for AI liveness detection.
    Returns: (confidence_score, status)
    """
    # Mock AI logic
    return 0.95, 'Real'
