requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Pillow
imagehash
Werkzeug
